+++ pigx-common/pigx-common-data/src/main/java/com/pig4cloud/pigx/common/data/datascope/DataScopeHandle.java
@@ -32,6 +32,6 @@ public interface DataScopeHandle {
 	 * @param deptList 部门ID，如果为空表示没有任何数据权限。
 	 * @return 返回true表示无需进行数据过滤处理，返回false表示需要进行数据过滤
 	 */
-	Boolean calcScope(List<Integer> deptList);
+	Boolean calcScope(List<Long> deptList);
 
 }
