+++ pigx-upms/pigx-upms-api/src/main/java/com/pig4cloud/pigx/admin/api/entity/SysDeptRelation.java
@@ -44,12 +44,12 @@ public class SysDeptRelation extends Model<SysDeptRelation> {
 	 * 祖先节点
 	 */
 	@ApiModelProperty(value = "祖先节点")
-	private Integer ancestor;
+	private Long ancestor;
 
 	/**
 	 * 后代节点
 	 */
 	@ApiModelProperty(value = "后代节点")
-	private Integer descendant;
+	private Long descendant;
 
 }
