+++ pigx-upms/pigx-upms-biz/src/main/java/com/pig4cloud/pigx/admin/mapper/SysRoleMapper.java
@@ -41,6 +41,6 @@ public interface SysRoleMapper extends PigxBaseMapper<SysRole> {
 	 * @param userId
 	 * @return
 	 */
-	List<SysRole> listRolesByUserId(Integer userId);
+	List<SysRole> listRolesByUserId(Long userId);
 
 }
