+++ pigx-upms/pigx-upms-biz/src/main/java/com/pig4cloud/pigx/admin/mapper/SysUserMapper.java
@@ -63,7 +63,7 @@ public interface SysUserMapper extends PigxBaseMapper<SysUser> {
 	 * @param id 用户ID
 	 * @return userVo
 	 */
-	UserVO getUserVoById(Integer id);
+	UserVO getUserVoById(Long id);
 
 	/**
 	 * 查询用户列表
