+++ pigx-upms/pigx-upms-biz/src/main/java/com/pig4cloud/pigx/admin/service/SysDeptService.java
@@ -39,7 +39,7 @@ public interface SysDeptService extends IService<SysDept> {
 	 * 查询部门树菜单
 	 * @return 树
 	 */
-	List<Tree<Integer>> selectTree();
+	List<Tree<Long>> selectTree();
 
 	/**
 	 * 添加信息部门
@@ -53,7 +53,7 @@ public interface SysDeptService extends IService<SysDept> {
 	 * @param id 部门 ID
 	 * @return 成功、失败
 	 */
-	Boolean removeDeptById(Integer id);
+	Boolean removeDeptById(Long id);
 
 	/**
 	 * 更新部门
